refactor layout to accommodate design request for…
… extend backgrounds all the way to the left and right, minus a bit of margin
… still ensure content doesn't run to wide within the full -width backgrounds.

in order to achieve that, we had to remove the container and margin on the body. this makes a few more classes / extends necessary, but saves us negative margins in a few places. Used calc() to ensure the max-width for the content, so we don't have to add extra container divs.

See general.scss for further comments.